fix: fixed the recipients viewing issue on touch screens
